pyproject.toml
[project]
name = "sql-data-visualisation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "psycopg2>=2.9.11",
    "python-telegram-bot>=22.7",
    "uvicorn>=0.42.0",
]
